FInd average assignment improves

+++ lessons/module1/les2.py
@@ -1,6 +1,6 @@
 def find_average():
     numbers = str(input("Enter 5 number, separated them with a comma: ")).split(", ")
-    numbers = [int(num) for num in numbers]
+    numbers = [float(num) for num in numbers]
     average = sum(numbers) / len(numbers) 
     return average
     # Uliana: Ik vind deze code wel goed, maar ik zou er ook een fouthandeling toegoeven om te checken wat gebruiker heeft ingetypt.
@@ -14,7 +14,6 @@ def calculate_age():
     return f"Your age: {age}"
     # Uliana: De code lijkt correct voor het berekenen van leeftijden, maar het zou beter zijn met ttoevoeging van foutafhandeling als currentYear < dan birthyear.
     # Ada: Ik vind de werking van deze code wel goede en het maakt goed zijn doel.
-
 
 def main():
     return calculate_age()
